# server/swagger.yaml
openapi: 3.0.0
info:
  title: Italian Restaurant API
  description: Docs for Italian Restaurant API.
  version: 1.0.0
host: localhost:5000
servers:
  - url: "/api/v1"
schemes:
  - http
...
paths:
...
  /sizes:
    get:
      security:
        - authToken: []
      tags:
        - size
      summary: Get all the product sizes [ADMIN]
      description: Get all the product sizes
      responses:
        "200": # status code
          description: A JSON array of product sizes
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  $ref: "#/components/schemas/Size"
        "401":
          description: Unauthorized error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsufficientPermissionsError"
    post:
      security:
        - authToken: []
      tags:
        - size
      summary: Create new size [ADMIN]
      description: Create new size
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Size"
            example:
              translations:
                ro:
                  value: "500 g"
                ru:
                  value: "500 г"
                en:
                  value: "500 g"
      responses:
        "200":
          description: "Product size was successfully created"
          schema:
            $ref: "#/components/schemas/Size"
        "422":
          description: "Size validation error"
          schema:
            $ref: "#/components/schemas/ValidationError"
        "404":
          description: "Resource not found"
          schema:
            $ref: "#/components/schemas/NotFoundError"
        "401":
          description: Unauthorized error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsufficientPermissionsError"

    put:
      security:
        - authToken: []
      tags:
        - size
      summary: Update a product size [ADMIN]
      description: Update a product size
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Size"
            example:
              _id: "62c9c25667eee57d19d4951e"
              translations:
                ro:
                  value: "500 g"
                ru:
                  value: "500 г"
                en: 
                  value: "500 g"
      responses:
        "200":
          description: "Size was successfully updated"
          schema:
            $ref: "#/components/schemas/Size"
        "422":
          description: "Size validation error"
          schema:
            $ref: "#/components/schemas/ValidationError"
        "404":
          description: "Resource not found"
          schema:
            $ref: "#/components/schemas/NotFoundError"
        "401":
          description: Unauthorized error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsufficientPermissionsError"
...
components:
  schemas:
...
    Size:
      type: object
      properties:
        _id: 
          type: string
        translations: 
          type: object
          properties: 
            ro: 
              type: object
              properties:
                value: 
                  type: string
            ru: 
              type: object
              properties:
                value: 
                  type: string
            en: 
              type: object
              properties:
                value: 
                  type: string
...
    ValidationError:
      type: object
      properties:
        status:
          type: number
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              key:
                type: string
    NotFoundError:
      type: object
      properties:
        status:
          type: number
        message:
          type: string
      example: 
        status: 404
        message: "RESOURCE_NOT_FOUND"
    UnauthorizedError:
      type: object
      properties:
        status:
          type: number
        message:
          type: string
      example:
        status: 401
        message: "NO_ACCESS_TOKEN"
    InsufficientPermissionsError:
      type: object
      properties:
        status:
          type: number
        message:
          type: string
      example:
        status: 403
        message: "INSUFFICIENT_PERMISSIONS"
  securitySchemes:
    authToken:
      name: "authToken"
      in: cookie
      type: apiKey
      description: "JWT authorization token sent in the authToken cookie."